requirements.txt
fastapi==0.95.1
uvicorn[standard]==0.22.0
python-dotenv==1.0.0
httpx==0.24.0
asyncpg==0.27.0
sqlalchemy==2.0.12
pydantic==1.10.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
redis==4.5.5
celery==5.2.7
spacy==3.5.3
transformers==4.29.2
torch==2.0.1
scikit-learn==1.2.2
pandas==2.0.1
numpy==1.24.3
matplotlib==3.7.1
python-dateutil==2.8.2 